requirements.txt
# Core dependencies
streamlit==1.22.0
python-dotenv==1.0.0
groq==0.18.0

# Document processing
pypdf2==3.0.1
python-docx==0.8.11

# Data processing and visualization
pandas==2.0.1
numpy==1.24.3
matplotlib==3.7.1
seaborn==0.12.2
wordcloud==1.9.2

# NLP libraries
spacy==3.6.0
scikit-learn==1.2.2

# Storage
pickle-mixin==1.0.2